// static-project/wsc-pc-vis/client/.yalc/@youzan/ebiz-components/components/easy-form/types.ts
export interface IFieldStatus {
  disabled: boolean;
  visible: boolean;
}
// ...
export type EasyFormWatchFunction<Value = any> = (
  value: Value,
  ctx: IEasyFormWatchCtx,
  easyForm: OriginalFormType
) => void;
// ...
interface IPartialFieldProps {
  className?: string;
  style?: React.CSSProperties;
  withoutLabel?: boolean;
}
// interface IEasyFormHooksLikeWatcher {
//   factory: EasyFormWatchFunction<Record<string, any>>;
//   deps: string[];
// }
export type EasyFormWatchType = /** IEasyFormHooksLikeWatcher |*/ Record<
  string,
  EasyFormWatchFunction<any>
>;
// ...
type EasyFormPlainType = {
  type: 'Plain';
  node: React.ReactNode;
} & IEasyFormBaseConfig;
export type EasyFormInternalListType = IForbidden_InternalListConfig & IEasyFormBaseConfig;
export type EasyFormInternalGroupType = IForbidden_InternalGroupConfig & IEasyFormBaseConfig;
export type EasyFormConfigValidType = (
  | IEasyFormInputConfig
  | IEasyFormNumberInputConfig
  | IEasyFormSelectConfig
  | IEasyFormCheckboxConfig
  | IEasyFormRadioConfig
  | IEasyFormCustomConfig
  | IEasyFormSwitchConfig
  | IEasyFormColorPickerConfig
  | IEasyFormDatePickerConfig
  | IEasyFormDateRangePickerConfig
) &
  IEasyFormBaseConfig;
export type EasyFormConfigType =
  | EasyFormPlainType
  | EasyFormConfigValidType
  | EasyFormInternalGroupType
  | EasyFormInternalListType;

export interface IGroupEasyFormConfigs {
  groupName: string;
  groupTitle: string | React.ReactNode;
  /**
   * 是否折叠分组
   */
  collapse?: boolean;
  config: EasyFormConfigType[];
  onChange?(value: Record<string, any>): void;
  /** 是否整块区域不可用，如果这个值被修改为false，则disabled属性会传到各个field上 */
  disabled?: boolean;
  watch?: EasyFormWatchType;
  hideGroupHeader?: boolean;
}

export interface IEasyFormListRepeatMethods {
  /** 添加一项，并将repeatConfig.defaultValue设置为初始值 */
  add(): void;
  /** 添加n个项，如果填写了`defaultValue`，就会将这个入参作为初始值来初始化 */
  add(addCount?: number, defaultValue?: any): void;
  /** 删除一项 */
  delete(): void;
  /** 删除n项，如果填写了`startIndex`，就会从`startIndex`开始删除n项 */
  delete(deleteCount?: number, startIndex?: number): void;
}

export interface IEasyFormRepeatTriggerProps {
  methods: IEasyFormListRepeatMethods;
  required: boolean;
  visible: boolean;
  disabled: boolean;
  label: string | null;
}
export interface IListEasyFormConfigs {
  listName: string;
  label?: string;
  fieldProps?: IPartialFieldProps & Record<string, any>;
  /**
   * 渲染触发数组变化的节点，该方法会带有几个实用方法，用于
   * 控制`repeatConfig`渲染的数量：
   * 1. `add`     添加一个或n个`repeatConfig`
   * 2. `delete`  删除最后一个或者删除第n个`repeatConfig`
   * 3. `filter`  过滤`repeatConfig`
   *
   * **请注意，一定要在返回的组件中，预留children的位置以供
   * 渲染list的内容**
   */
  repeatTrigger: React.ComponentType<IEasyFormRepeatTriggerProps>;
  /**
   * 需要重复的field配置，**注意：在`list`中，因为所有数据列
   * 表都是通过这个属性来渲染的，所以不允许使用`watch`方法来监
   * 听其他数据。**
   */
  repeatConfig: EasyFormConfigType;
  /**
   * 在重复生成`repeatConfig`的时候，需要指定唯一id
   * 如果没有声明这个方法，会采用`name + index`的方式
   * 命名每一个key
   */
  getRepeatKey?(name: string, idx: number): any;
  onChange?(value: any[]): void;
  /** 如果这个值为false，则表示包括，trigger在内都不会显示 */
  visible?: boolean;
  /** 这个参数将会影响repeatTrigger的disabled参数 */
  disabled?: boolean;
  watch?: EasyFormWatchType;
}

interface IFieldBaseConfigHandlerProps<ValueType> {
  defaultValue?: ValueType;
  onChange?(value: ValueType): void;
  onBlur?(value: ValueType): void;
}
export type FieldInjectPropsType<ValueType> = IEasyFormBaseConfig &
  Required<IFieldBaseConfigHandlerProps<ValueType>>;

export interface IEasyFormFieldRenderProps<Model, StatusModel, Config = EasyFormConfigType> {
  model: Model;
  statusModel: StatusModel;
  config: Config;
  addColon: boolean;
  disabled?: boolean;
  onChange(key: string, value: any): void;
}

interface IEasyFormFiledBaseProps {
  value: any;
  onChange(input: any): void;
}
export type EasyFormFieldProps<P> = IEasyFormFiledBaseProps & P;
/** 内部类型的数据格式，比如list和group类型，需要有自己的状态，需要额外定义成set类型 */
export type EasyFormInternalType<P extends BasicModel<any>> = FieldSetModel<{
  $$type: FieldModel<string>;
  status: FieldModel<Record<string, any>>;
  value: P;
}>;

// model types
export type NormalEasyFormModel = FieldModel<any>;
export type GroupEasyFormModel = FieldSetModel<Record<string, NormalEasyFormModel>>;
export type ListEasyFormModel = FieldArrayModel<any>;
export type EasyFormModelType = NormalEasyFormModel | GroupEasyFormModel | ListEasyFormModel;
// status type
export type NormalStatusModel = FieldModel<IFieldStatus>;
export type GroupStatusModel = FieldSetModel<Record<string, NormalStatusModel>>;
export type ListStatusModel = NormalStatusModel;